Abstract DFV API and start abstracting config handling
